sre-agents/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
python-dotenv>=1.0.0
watchdog>=4.0.0
flask>=3.0.0
